Fixes #78 date_end gets reset to date_start
Hobby models had a bug in save() whereby the condition to check if the starting date was after the ending date was checked the other way round. This caused the ending date to be reset to the starting date everytime the model was saved. Added tests to ensure that the save method works as expected.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>