Fixes #82 feedback_add url does not work in prod
The feedback_add url has an optional parameter for passing urls that can be blank. The misconfigured url regex required there to be two trailing slashes at the end, which caused the 404 error in prod. Fixed by omitting the trailing slash since the url itself contains a trailing slash when given.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>