fixes #84 lifeX post save accepts duplicates
LifeXPost.save() accepted duplicates via same idea and week values. This resulted in the same idea being tried more than once in a single week, and resulted in an error in GET requests for that particular posts.
This has been fixed by asserting that no other posts with the same idea and week exist before saving the current post.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>