Pretty code and Pagedown editor for devblog posts
Use django-pagedown for devblog post admin and markdown-deux to render the markdown in template. Use google code-prettify to format the code according to language.

Code prettify requires code blocks to be tagged with the `prettify` class, which cannot be done in markdown. A jquery line has been added to automatically supplement all `pre` blocks with the `prettify` class.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>