change:make slug field optional in admin
Add option blank=True to make the slug field optional in the admin since it is automatically generated in the save method. Changed the date_created, date_updated to make date_created an editable field, and date_updated to be automatically set every time the object is saved. Added verbose name to is_published.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>