fix: D3.js for lifeX:goals load error over HTTP
D3.js which is a requirement for showing lifeX goals can no longer be loaded using HTTP due to the browser's security policy of loading non-secure content (site is now HTTPS). URLs generated by django are not appended with HTTPS, and are not redirected in a JS request. Fixed this by replacing the HTTP with HTTPS in javascript url.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>