fix: add None as a subdomain to sitebase footer
Sitebase footer contains links to items on the base site, i.e. those not under any subdomains. Without an explicit subdomain, these are written with the current subdomain. This is fixed by explicitly using the subdomain = None, which reverts it to either www, or None, as specified in the urlconf.

Signed-off-by: Harshvardhan Pandit <coolharsh55@gmail.com>